pyproject.toml
[tool.poetry]
name = "tornado_imageboard"
version = "0.1.1"
description = "Python (Tornado) based imageboard engine."
authors = ["cnsr <dmytrenkoyar@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
tornado = "^6.1.0"
Pillow = "^8.1.2"
bcrypt = "^3.2.0"
certifi = "^2019.9.11"
cffi = "^1.13.1"
chardet = "^3.0.4"
geoip2 = "^2.9.0"
idna = "^2.8"
itsdangerous = "^1.1.0"
maxminddb = "^1.5.1"
motor = "^2.0.0"
olefile = "^0.46"
pipe = "^1.5.0"
popen = "^0.1.20"
pycparser = "^2.19"
pymongo = "^3.9.0"
requests = "^2.25.1"
six = "^1.15.0"
SQLAlchemy = "^1.3.10"
tornado-sqlalchemy = "^0.6.1"
urllib3 = "^1.25.11"
isort = "^5.8.0"
python-dotenv = "^0.15.0"
PyJWT = "^2.1.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.0"
black = "^20.8b1"
mypy = "^0.812"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
